requirements.txt
numpy
pyyaml
cryptography
rich
psutil
toml
networkx
z3-solver
matplotlib
pysmt
pretty-errors
shapely
tqdm
pandas == 2.2.2
yaspin
scikit-learn == 1.3.2 
tensorflow == 2.15.0
joblib
cryptography
rich
psutil
pyyaml
yaspin
sqlalchemy==1.4.16
torch == 2.3.1
boto3
grpcio
grpcio-tools
python-dotenv==1.0.1
torch_geometric==2.3.0
optuna==4.0.0
optuna-integration==4.0.0
mlflow==2.17.2
fpdf2==2.8.1
pyinstaller==6.11.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
idna==3.10
proto-plus==1.26.1
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests==2.32.3
rsa==4.9.1
urllib3==2.4.0
gdsast==1.0.2
elasticsearch==8.12.0